Initial commit
This is taking the freeswitch softphone example and wrapping it in a
javascript object.  The goal is to fully encasulate the phone part from a
ui.  Currently restricts itself to one registration and one call.  Uses
custom javascript events for awesomeness.

The freeswitch.mxml was modified to allow for a suppressed ring sound so
something else can ring on the onIncomingCall event.